+++ src/main/java/rabun/oanda/rest/models/OrderMarket.java
@@ -1,7 +1,7 @@
 package rabun.oanda.rest.models;
 
 public class OrderMarket extends Order {
-    public int units;
+    public long units;
     public float takeProfit;
     public float stopLoss;
     public float trailingStop;
