+++ src/main/java/rabun/oanda/rest/models/Position.java
@@ -2,7 +2,7 @@ package rabun.oanda.rest.models;
 
 public class Position {
     public String instrument;
-    public int units;
+    public long units;
     public OandaTypes.Side side;
     public float avgPrice;
 }
