+++ src/main/java/rabun/oanda/rest/models/PositionClosed.java
@@ -3,6 +3,6 @@ package rabun.oanda.rest.models;
 public class PositionClosed {
     public int[] ids;
     public String instrument;
-    public int totalUnits;
+    public long totalUnits;
     public float price;
 }
