+++ src/main/java/rabun/oanda/rest/models/TransactionMigrateTradeClose.java
@@ -2,7 +2,7 @@ package rabun.oanda.rest.models;
 
 public class TransactionMigrateTradeClose extends TransactionSimple {
     public float pl;
-    public int interest;
+    public long interest;
     public float accountBalance;
-    public int tradeId;
+    public long tradeId;
 }
