+++ src/main/java/rabun/oanda/rest/models/TransactionOrderCancel.java
@@ -2,6 +2,6 @@ package rabun.oanda.rest.models;
 
 
 public class TransactionOrderCancel extends Transaction {
-    public int orderId;
+    public long orderId;
     public OandaTypes.Reason reason;
 }
