+++ src/main/java/rabun/oanda/rest/models/TransactionTakeProfitField.java
@@ -1,8 +1,8 @@
 package rabun.oanda.rest.models;
 
 public class TransactionTakeProfitField extends TransactionSimple {
-    public int tradeId;
+    public long tradeId;
     public float pl;
-    public int interest;
+    public long interest;
     public float accountBalance;
 }
